requirements.txt
attrs>=22.1.0
GitPython
tqdm
joblib
